LayoutTests/platform/gtk/imported/w3c/web-platform-tests/navigation-timing/test_timing_xserver_redirect-expected.txt

Description

This test validates the values of the window.performance.redirectCount and the window.performance.timing.redirectStart/End times for a cross-origin server side redirect navigation.


PASS window.performance is defined
PASS window.performance.navigation is defined
PASS timing.navigation.type is TYPE_NAVIGATE
FAIL navigation.redirectCount == 0 on a cross-origin server redirected navigation assert_equals: navigation.redirectCount == 0 on a cross-origin server redirected navigation expected 0 but got 1
PASS window.performance.timing.navigationStart > 0
FAIL timing.redirectStart == 0 on a server redirected navigation from another domain assert_equals: timing.redirectStart == 0 on a server redirected navigation from another domain expected 0 but got 1675047620791
FAIL timing.redirectEnd == 0 on a server redirected navigation from another domain assert_equals: timing.redirectEnd == 0 on a server redirected navigation from another domain expected 0 but got 1675047620802
